AsWixSchema to types
