+++ libs/external-db-postgres/src/postgres_operations.ts
@@ -9,6 +9,7 @@ export default class DatabaseOperations implements IDatabaseOperations {
         this.pool = pool
     }
 
+    // @ts-ignore
     async validateConnection() {
         return await this.pool.query('SELECT 1').then(() => { return { valid: true } })
                               .catch((e) => { return { valid: false, error: notThrowingTranslateErrorCodes(e) } })
