+++ libs/velo-external-db-commons/src/libs/types.ts
@@ -7,18 +7,3 @@ export type Field = {
     precision?: number,
     isPrimary?: boolean,
 }
-
-export type FieldWithQueryOperators = ResponseField & { queryOperators: string[] }
-
-export interface AsWixSchemaHeaders {
-    id: string,
-    displayName: string,
-    maxPageSize: number,
-    ttl: number
-}
-
-export interface AsWixSchema extends AsWixSchemaHeaders {
-    allowedOperations: string[],
-    allowedSchemaOperations: string[],
-    fields: { [field: string]: FieldWithQueryOperators }
-}
